+++ app/resolvers/default/MachineResolver.js
@@ -67,9 +67,9 @@ function MachineResolver({machineService, saleService, kktService}) {
 
     const salesSummary = async (obj, args, context) => {
         const {user} = context
-        const {period} = args
+        const {period, machineGroupId} = args
 
-        const salesSummary = await saleService.getSalesSummary({machineId: obj.id, period}, user)
+        const salesSummary = await saleService.getSalesSummary({machineGroupId, machineId: obj.id, period}, user)
 
         if (!salesSummary) {
             return null
@@ -94,12 +94,13 @@ function MachineResolver({machineService, saleService, kktService}) {
 
     const salesByEncashment = async (obj, args, context) => {
         const {user} = context
+        const {machineGroupId} = args
 
         const lastEncashment = await machineService.getLastMachineEncashment(obj.id, user)
 
         const period = {from: lastEncashment ? lastEncashment.timestamp : new Date(0), to: new Date()}
 
-        const salesSummary = await saleService.getSalesSummary({machineId: obj.id, period}, user)
+        const salesSummary = await saleService.getSalesSummary({machineGroupId, machineId: obj.id, period}, user)
 
         if (!salesSummary) {
             return null